Display x : x2 => long double

+++ src/tools.c
@@ -40,12 +40,12 @@ void display_division_zero(void)
 
 void display_x_value(double x, config_t *config)
 {
-    double x2;
+    long double x2;
 
-    x = round(x * pow(10, config->precision + 5)) / pow(10, config->precision + 5);
-    x2 = x * pow(10, config->precision);
+    x = round(x * powl(10, config->precision + 5)) / powl(10, config->precision + 5);
+    x2 = x * powl(10, config->precision);
     x2 = x2 - ((int)x2);
-    x2 *= pow(10, 5);
+    x2 *= powl(10, 5);
     x2 = floor(x2);
     if (x2 != 0) {
         printf("x = %.*f\n", get_nb_decimal(x, config->precision), x);
